+++ plugins/pmsmart/src/main/java/org/itsallcode/whiterabbit/plugin/pmsmart/web/page/WeekViewPage.java
@@ -8,7 +8,6 @@ import java.time.format.DateTimeFormatter;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.itsallcode.whiterabbit.plugin.pmsmart.web.Driver;
-import org.itsallcode.whiterabbit.plugin.pmsmart.web.Element;
 import org.openqa.selenium.By;
 
 public class WeekViewPage implements Page
@@ -40,8 +39,7 @@ public class WeekViewPage implements Page
         LOG.debug("Selecting week for day {}...", day);
 
         driver.findElement(By.id("MainContent_EdtDate_B-1")).click();
-        final DateSelector dateSelector = new DateSelector(driver,
-                driver.findElement(By.id("MainContent_EdtDate_DDD_PW-1")));
+        final var dateSelector = new DateSelector(driver, driver.findElement(By.id("MainContent_EdtDate_DDD_PW-1")));
         dateSelector.select(day);
         if (!isDaySelected(day))
         {
@@ -63,19 +61,19 @@ public class WeekViewPage implements Page
 
     public ProjectTable getProjectTable()
     {
-        final Element table = driver
+        final var table = driver
                 .findElement(By.xpath("//*[@id=\"MainContent_ASPxCpWbGrid_WeekBookingGrid_DXMainTable\"]"));
         return new ProjectTable(driver, table, this);
     }
 
     public LocalDate getSelectedWeekFirstDay()
     {
-        final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd.MM.yyyy");
+        final var formatter = DateTimeFormatter.ofPattern("dd.MM.yyyy");
         final String text = getCurrentWeek();
         final String[] parts = text.split("[\\s-,]");
 
-        final LocalDate firstDay = LocalDate.parse(parts[0], formatter);
-        final LocalDate lastDay = LocalDate.parse(parts[3], formatter);
+        final var firstDay = LocalDate.parse(parts[0], formatter);
+        final var lastDay = LocalDate.parse(parts[3], formatter);
         LOG.trace("First day: {}, last day: {}", firstDay, lastDay);
         return firstDay;
     }